+++ e2e/nextjs-routes.d.ts
@@ -2,26 +2,23 @@
 // Run `yarn nextjs-routes` to regenerate this file.
 /* eslint-disable */
 
-type Route =
-  | { pathname?: "/foos/[foo]"; query: Query<{ foo: string }> }
-  | { pathname?: "/"; query?: Query | undefined };
+declare module "nextjs-routes" {
+  export type Route =
+    | { pathname: "/foos/[foo]"; query: Query<{ foo: string }> }
+    | { pathname: "/"; query?: Query | undefined };
 
-type Query<Params = {}> = Params & {
-  [key: string]: string;
-};
-
-type Pathname = Exclude<Route["pathname"], undefined>;
-
-type QueryForPathname = {
-  [K in Route as K["pathname"]]: Exclude<K["query"], undefined>;
-};
+  type Query<Params = {}> = Params & { [key: string]: string | undefined };
+}
 
 declare module "next/link" {
+  import type { Route } from "nextjs-routes";
   import type { LinkProps as NextLinkProps } from "next/dist/client/link";
   import type { PropsWithChildren, MouseEventHandler } from "react";
 
+  type RouteOrQuery = Route | { query?: { [key: string]: string | undefined } };
+
   export interface LinkProps extends Omit<NextLinkProps, "href"> {
-    href: Route;
+    href: RouteOrQuery;
   }
 
   declare function Link(
@@ -40,10 +37,19 @@ declare module "next/link" {
 }
 
 declare module "next/router" {
+  import type { Route } from "nextjs-routes";
   import type { NextRouter as Router } from "next/dist/client/router";
   export { RouterEvent } from "next/dist/client/router";
 
   type TransitionOptions = Parameters<Router["push"]>[2];
+
+  type Pathname = Route["pathname"];
+
+  type QueryForPathname = {
+    [K in Route as K["pathname"]]: Exclude<K["query"], undefined>;
+  };
+
+  type RouteOrQuery = Route | { query: { [key: string]: string | undefined } };
 
   export interface NextRouter<P extends Pathname = Pathname>
     extends Omit<Router, "push" | "replace"> {
@@ -51,12 +57,12 @@ declare module "next/router" {
     route: P;
     query: QueryForPathname[P];
     push(
-      url: Route,
+      url: RouteOrQuery,
       as?: string,
       options?: TransitionOptions
     ): Promise<boolean>;
     replace(
-      url: Route,
+      url: RouteOrQuery,
       as?: string,
       options?: TransitionOptions
     ): Promise<boolean>;